requirements.txt
diffusers
transformers
accelerate
peft
gradio
opencv-python
numpy
pillow
